+++ app/components/annotations/annotations.js
@@ -1,0 +1,6 @@
+/**
+ * @module
+ * @description Annotations module shows an assets annotations and adds support
+ *              for adding annotations to an asset.
+ */
+angular.module('annotations', ['ngResource']);